requirements.txt
streamlit
duckdb
pandas
pyarrow
python-dotenv
yfinance